+++ src/gsply/__init__.py
@@ -68,7 +68,7 @@ from gsply.utils import (
 )
 from gsply.writer import compress_to_arrays, compress_to_bytes, plywrite
 
-__version__ = "0.4.5"
+__version__ = "0.4.6"
 __all__ = [
     "plyread",
     "GSData",